requirements.txt
torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2
xformers==0.0.25
pytorch-lightning==2.2.1
transformers==4.40.0
einops==0.6.1
timm==0.6.13
gradio==3.16.2
albumentations==1.3.0
opencv-contrib-python==4.7.0.72
imageio==2.9.0
imageio-ffmpeg==0.4.2
omegaconf==2.1.1
test-tube>=0.7.5
streamlit==1.12.1
webdataset==0.2.5
kornia==0.6
numpy==1.26.4
invisible-watermark>=0.1.5
streamlit-drawable-canvas==0.8.0
torchmetrics==1.3.0
addict==2.4.0
yapf==0.32.0
prettytable==3.6.0
safetensors==0.4.2

# DINOv3 相关依赖
dinov3 @ git+https://github.com/facebookresearch/dinov3.git
open-clip-torch==2.0.2
Pillow>=9.0.0
scikit-image>=0.19.0
matplotlib>=3.5.0

